+++ src/commands/cd.js
@@ -9,6 +9,7 @@ const Verbose = require("../classes/Verbose");
 const InfoMessages = require("../classes/InfoMessages");
 const PathUtil = require("../classes/PathUtil");
 const SettingManager = require("../classes/SettingManager");
+const ConfigManager = require("../classes/ConfigManager");
 
 /**
  * The `cd` command, used to change into a directory.
@@ -38,6 +39,41 @@ const cd = (dir, ...args) => {
       Errors.enterParameter("a directory", "cd test");
       return;
     }
+
+    const config = new ConfigManager();
+
+    if (dir.endsWith("-")) {
+      const lastDir = config.getConfig()?.cd?.lastDir;
+
+      if (typeof lastDir === "undefined") {
+        InfoMessages.error("Could not get previous directory.");
+        return;
+      }
+
+      // Retrieve the second-to-last directory, if available
+      const secondLastDir = config.getConfig()?.cd?.secondLastDir;
+
+      if (typeof secondLastDir === "undefined") {
+        InfoMessages.error("Could not get second last directory.");
+        return;
+      }
+
+      // Change to the second last directory
+      process.chdir(secondLastDir);
+      if (!silent)
+        InfoMessages.success(
+          `Successfully changed the directory to ${chalk.bold(
+            new SettingManager().fullOrBase(secondLastDir)
+          )}.`
+        );
+      else console.log();
+      return;
+    }
+
+    // Update the config to save the current and previous directories
+    const currentDir = dir;
+    const previousDir = config.getConfig()?.cd?.lastDir || currentDir;
+    config.addData({ cd: { lastDir: currentDir, secondLastDir: previousDir } });
 
     if (!dirChk.doesExist()) {
       Verbose.chkExists(dir);