+++ src/functions/init/initConfig.js
@@ -1,29 +1,30 @@
 const settings = require("../../data/settings.json");
 
+const { BUILD } = require("../../variables/constants");
+
 const ConfigManager = require("../../classes/ConfigManager");
 
 const _initConfig = () => {
   const config = new ConfigManager();
+  let errorEncountered = false;
 
-  // Reset config file
-  config.deleteConfig();
-  config.createConfig();
+  const actions = [
+    config.deleteConfig(),
+    config.createConfig(),
+    config.addData({
+      settings: Object.fromEntries(
+        Object.entries(settings).map(([key, setting]) => [
+          key,
+          { displayName: setting.displayName, current: setting.default },
+        ])
+      ),
+      history: [],
+      build: BUILD,
+    }),
+  ];
 
-  // Add default settings
-  config.addData({
-    settings: Object.fromEntries(
-      Object.entries(settings).map(([key, setting]) => [
-        key,
-        {
-          displayName: setting.displayName,
-          current: setting.default,
-        },
-      ])
-    ),
-  });
-
-  // Add history key to prevent config corruption error
-  config.addData({ history: [] });
+  errorEncountered = actions.some((action) => !action);
+  return errorEncountered;
 };
 
 module.exports = _initConfig;